# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Finance App API",
        "description": "REST API for personal finance management built with Node.js, Express, Prisma and PostgreSQL.\n\n## Features\n- JWT authentication (access + refresh tokens)\n- User profile under `/api/users/me`\n- Financial transactions with pagination, filters and safe sorting\n- Standardized responses: `{ success, message, data }` and errors `{ success, message, code }`\n- Rate limiting, CORS, Helmet and ownership checks on protected resources\n\n## Authentication\nProtected routes require the header:\n`Authorization: Bearer <accessToken>`\n\nObtain tokens via `POST /api/users` (register) or `POST /api/users/login`. Use `POST /api/users/refresh-token` to renew tokens.\n\n## Notes\n- Passwords are never returned in API responses.\n- `user_id` in request bodies is ignored; the authenticated user from the JWT is always used.\n- Unsupported query parameters (e.g. `categoryId`) return `400 VALIDATION_ERROR`.",
        "version": "1.0.0",
        "contact": {
            "name": "Finance App",
            "url": "https://github.com/vivianeaguiarc/finance-app"
        }
    },
    "host": "finance-app-i600.onrender.com",
    "basePath": "/",
    "schemes": ["https", "http"],
...
    "securityDefinitions": {
        "bearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header",
            "description": "JWT access token. Format: `Authorization: Bearer <accessToken>`"
        }
    },
    "paths": {
...
        "/api/transactions/me/installments": {
            "post": {
                "tags": ["Transactions"],
                "summary": "Create installment purchase",
                "description": "Example: R$ 1200 in 12x generates 12 transactions of R$ 100 atomically.",
                "security": [{ "bearerAuth": [] }],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "required": [
                                "name",
                                "date",
                                "type",
                                "totalAmount",
                                "totalInstallments"
                            ],
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "example": "Notebook"
                                },
                                "date": {
                                    "type": "string",
                                    "format": "date-time"
                                },
                                "type": {
                                    "type": "string",
                                    "enum": ["EXPENSE", "EARNING", "INVESTMENT"]
                                },
                                "totalAmount": {
                                    "type": "number",
                                    "example": 1200
                                },
                                "totalInstallments": {
                                    "type": "integer",
                                    "example": 12
                                },
                                "categoryId": {
                                    "type": "string",
                                    "format": "uuid"
                                }
                            }
                        }
                    }
                ],
                "responses": { "201": { "description": "Created" } }
            }
        },
        "/api/transactions/me/recurring": {
            "post": {
                "tags": ["Transactions"],
                "summary": "Create recurring transaction series",
                "description": "Example: monthly rent/salary until recurrenceEndDate.",
                "security": [{ "bearerAuth": [] }],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "required": [
                                "name",
                                "date",
                                "type",
                                "amount",
                                "isRecurring",
                                "recurrenceType"
                            ],
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "example": "Rent"
                                },
                                "date": {
                                    "type": "string",
                                    "format": "date-time"
                                },
                                "type": {
                                    "type": "string",
                                    "enum": ["EXPENSE", "EARNING", "INVESTMENT"]
                                },
                                "amount": { "type": "number", "example": 800 },
                                "isRecurring": {
                                    "type": "boolean",
                                    "example": true
                                },
                                "recurrenceType": {
                                    "type": "string",
                                    "enum": ["WEEKLY", "MONTHLY", "YEARLY"]
                                },
                                "recurrenceEndDate": {
                                    "type": "string",
                                    "format": "date-time"
                                },
                                "categoryId": {
                                    "type": "string",
                                    "format": "uuid"
                                }
                            }
                        }
                    }
                ],
                "responses": { "201": { "description": "Created" } }
            },
            "get": {
                "tags": ["Transactions"],
                "summary": "List recurring transaction templates",
                "security": [{ "bearerAuth": [] }],
                "responses": { "200": { "description": "OK" } }
            }
        },
        "/api/transactions/me": {
            "get": {
                "tags": ["Transactions"],
                "summary": "List my transactions (paginated)",
                "description": "Returns paginated transactions for the authenticated user. Supports filters and allowlisted sorting. Unknown query params (e.g. `categoryId`) return 400.",
                "security": [{ "bearerAuth": [] }],
                "parameters": [
                    { "$ref": "#/parameters/PageParam" },
                    { "$ref": "#/parameters/LimitParam" },
                    { "$ref": "#/parameters/TypeParam" },
                    { "$ref": "#/parameters/StartDateParam" },
                    { "$ref": "#/parameters/EndDateParam" },
                    { "$ref": "#/parameters/FromParam" },
                    { "$ref": "#/parameters/ToParam" },
                    { "$ref": "#/parameters/MinAmountParam" },
                    { "$ref": "#/parameters/MaxAmountParam" },
                    { "$ref": "#/parameters/SortByParam" },
                    { "$ref": "#/parameters/SortOrderParam" },
                    { "$ref": "#/parameters/CategoryIdParam" }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list",
                        "schema": {
                            "$ref": "#/definitions/PaginatedTransactionResponse"
                        },
                        "examples": {
                            "application/json": {
                                "success": true,
                                "message": "Transactions retrieved successfully",
                                "data": [
                                    {
                                        "id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                                        "user_id": "7c9e6679-7425-40de-944b-e07fc1f90ae7",
                                        "name": "Groceries",
                                        "type": "EXPENSE",
                                        "amount": "150.00",
                                        "date": "2024-06-01T00:00:00.000Z",
                                        "created_at": "2024-06-01T12:00:00.000Z"
                                    }
                                ],
                                "meta": {
                                    "page": 1,
                                    "limit": 10,
                                    "total": 42,
                                    "totalPages": 5
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid query parameters",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        },
                        "examples": {
                            "application/json": {
                                "success": false,
                                "message": "Unallowed query parameter",
                                "code": "VALIDATION_ERROR"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Rate limit exceeded",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "tags": ["Transactions"],
                "summary": "Create a transaction",
                "description": "Creates a transaction for the authenticated user. Any `user_id` sent in the body is ignored.",
                "security": [{ "bearerAuth": [] }],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateTransactionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Transaction created",
                        "schema": {
                            "$ref": "#/definitions/TransactionSuccessResponse"
                        },
                        "examples": {
                            "application/json": {
                                "success": true,
                                "message": "Transaction created successfully",
                                "data": {
                                    "id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                                    "user_id": "7c9e6679-7425-40de-944b-e07fc1f90ae7",
                                    "name": "Salary",
                                    "type": "EARNING",
                                    "amount": "5000.00",
                                    "date": "2024-06-01T00:00:00.000Z",
                                    "created_at": "2024-06-01T12:00:00.000Z"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/transactions": {
            "get": {
                "tags": ["Transactions"],
                "summary": "List transactions (alias)",
                "description": "Alias of `GET /api/transactions/me` with the same pagination, filters and security.",
                "security": [{ "bearerAuth": [] }],
                "parameters": [
                    { "$ref": "#/parameters/PageParam" },
                    { "$ref": "#/parameters/LimitParam" },
                    { "$ref": "#/parameters/TypeParam" },
                    { "$ref": "#/parameters/StartDateParam" },
                    { "$ref": "#/parameters/EndDateParam" },
                    { "$ref": "#/parameters/FromParam" },
                    { "$ref": "#/parameters/ToParam" },
                    { "$ref": "#/parameters/MinAmountParam" },
                    { "$ref": "#/parameters/MaxAmountParam" },
                    { "$ref": "#/parameters/SortByParam" },
                    { "$ref": "#/parameters/SortOrderParam" },
                    { "$ref": "#/parameters/CategoryIdParam" }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list",
                        "schema": {
                            "$ref": "#/definitions/PaginatedTransactionResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid query parameters",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "tags": ["Transactions"],
                "summary": "Create transaction (alias)",
                "description": "Alias of `POST /api/transactions/me`.",
                "security": [{ "bearerAuth": [] }],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateTransactionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Transaction created",
                        "schema": {
                            "$ref": "#/definitions/TransactionSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/transactions/me/{transactionId}": {
            "patch": {
                "tags": ["Transactions"],
                "summary": "Update a transaction",
                "description": "Updates a transaction owned by the authenticated user. Returns 403 if the transaction belongs to another user.",
                "security": [{ "bearerAuth": [] }],
                "parameters": [
                    {
                        "name": "transactionId",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "format": "uuid",
                        "description": "Transaction UUID"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdateTransactionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction updated",
                        "schema": {
                            "$ref": "#/definitions/TransactionSuccessResponse"
                        },
                        "examples": {
                            "application/json": {
                                "success": true,
                                "message": "Transaction updated successfully",
                                "data": {
                                    "id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                                    "user_id": "7c9e6679-7425-40de-944b-e07fc1f90ae7",
                                    "name": "Updated salary",
                                    "type": "EARNING",
                                    "amount": "5500.00",
                                    "date": "2024-06-15T00:00:00.000Z",
                                    "created_at": "2024-06-01T12:00:00.000Z"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden — not the owner",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        },
                        "examples": {
                            "application/json": {
                                "success": false,
                                "message": "You do not have permission to perform this action.",
                                "code": "FORBIDDEN"
                            }
                        }
                    },
                    "404": {
                        "description": "Transaction not found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        },
                        "examples": {
                            "application/json": {
                                "success": false,
                                "message": "Transaction not found.",
                                "code": "TRANSACTION_NOT_FOUND"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "tags": ["Transactions"],
                "summary": "Delete a transaction",
                "description": "Deletes a transaction owned by the authenticated user.",
                "security": [{ "bearerAuth": [] }],
                "parameters": [
                    {
                        "name": "transactionId",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "format": "uuid",
                        "description": "Transaction UUID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction deleted",
                        "schema": {
                            "$ref": "#/definitions/TransactionSuccessResponse"
                        },
                        "examples": {
                            "application/json": {
                                "success": true,
                                "message": "Transaction deleted successfully",
                                "data": {
                                    "id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                                    "user_id": "7c9e6679-7425-40de-944b-e07fc1f90ae7",
                                    "name": "Groceries",
                                    "type": "EXPENSE",
                                    "amount": "150.00",
                                    "date": "2024-06-01T00:00:00.000Z",
                                    "created_at": "2024-06-01T12:00:00.000Z"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden — not the owner",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Transaction not found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "parameters": {
        "PageParam": {
            "name": "page",
            "in": "query",
            "required": false,
            "type": "integer",
            "default": 1,
            "minimum": 1,
            "description": "Page number (default: 1)"
        },
        "LimitParam": {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "integer",
            "default": 10,
            "minimum": 1,
            "maximum": 100,
            "description": "Items per page (default: 10, max: 100)"
        },
        "TypeParam": {
            "name": "type",
            "in": "query",
            "required": false,
            "type": "string",
            "enum": ["EARNING", "EXPENSE", "INVESTMENT"],
            "description": "Filter by transaction type"
        },
        "StartDateParam": {
            "name": "startDate",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date-time",
            "description": "Inclusive start of date range (ISO 8601)",
            "example": "2024-01-01T00:00:00.000Z"
        },
        "EndDateParam": {
            "name": "endDate",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date-time",
            "description": "Inclusive end of date range (ISO 8601)",
            "example": "2024-12-31T23:59:59.999Z"
        },
        "FromParam": {
            "name": "from",
            "in": "query",
            "required": false,
            "type": "string",
            "description": "Legacy alias for startDate",
            "example": "2024-01-01"
        },
        "ToParam": {
            "name": "to",
            "in": "query",
            "required": false,
            "type": "string",
            "description": "Legacy alias for endDate",
            "example": "2024-12-31"
        },
        "MinAmountParam": {
            "name": "minAmount",
            "in": "query",
            "required": false,
            "type": "number",
            "minimum": 0,
            "description": "Minimum transaction amount"
        },
        "MaxAmountParam": {
            "name": "maxAmount",
            "in": "query",
            "required": false,
            "type": "number",
            "minimum": 0,
            "description": "Maximum transaction amount"
        },
        "SortByParam": {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "type": "string",
            "enum": ["date", "amount", "createdAt"],
            "default": "date",
            "description": "Sort field (allowlist only)"
        },
        "SortOrderParam": {
            "name": "sortOrder",
            "in": "query",
            "required": false,
            "type": "string",
            "enum": ["asc", "desc"],
            "default": "desc",
            "description": "Sort direction. Use `sortOrder` (not `order`)."
        },
        "CategoryIdParam": {
            "name": "categoryId",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "uuid",
            "description": "Not supported on transaction list — returns 400. Use `/api/dashboard` with category name filter instead."
        },
...
    },
    "definitions": {
...
        "ErrorResponse": {
            "type": "object",
            "required": ["success", "message", "code"],
            "properties": {
                "success": {
                    "type": "boolean",
                    "example": false
                },
                "message": {
                    "type": "string",
                    "example": "Resource not found."
                },
                "code": {
                    "type": "string",
                    "example": "NOT_FOUND",
                    "description": "Machine-readable error code (e.g. VALIDATION_ERROR, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, EMAIL_ALREADY_IN_USE, TOO_MANY_REQUESTS)"
                }
            }
        },
...
        "TransactionSuccessResponse": {
            "type": "object",
            "properties": {
                "success": { "type": "boolean", "example": true },
                "message": { "type": "string" },
                "data": {
                    "$ref": "#/definitions/TransactionResponse"
                }
            }
        },
        "PaginationMeta": {
            "type": "object",
            "properties": {
                "page": { "type": "integer", "example": 1 },
                "limit": { "type": "integer", "example": 10 },
                "total": { "type": "integer", "example": 42 },
                "totalPages": { "type": "integer", "example": 5 }
            }
        },
...
        "PaginatedTransactionResponse": {
            "type": "object",
            "properties": {
                "success": { "type": "boolean", "example": true },
                "message": {
                    "type": "string",
                    "example": "Transactions retrieved successfully"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/TransactionResponse"
                    }
                },
                "meta": {
                    "$ref": "#/definitions/PaginationMeta"
                }
            }
        },
...
        "CreateTransactionRequest": {
            "type": "object",
            "required": ["name", "type", "date", "amount"],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Salary"
                },
                "type": {
                    "type": "string",
                    "enum": ["EARNING", "EXPENSE", "INVESTMENT"],
                    "example": "EARNING"
                },
                "date": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2024-06-01T00:00:00.000Z"
                },
                "amount": {
                    "type": "number",
                    "minimum": 1,
                    "example": 5000
                }
            }
        },
        "UpdateTransactionRequest": {
            "type": "object",
            "properties": {
                "name": { "type": "string", "example": "Updated salary" },
                "type": {
                    "type": "string",
                    "enum": ["EARNING", "EXPENSE", "INVESTMENT"]
                },
                "date": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2024-06-15T00:00:00.000Z"
                },
                "amount": {
                    "type": "number",
                    "minimum": 1,
                    "example": 5500
                }
            }
        }
    }
}